🚚 Move settings-tab files to correct folder
