+++ modules/ppcp-settings/resources/js/Components/Screens/Settings/Components/Settings/Blocks/HooksListBlock.js
@@ -4,7 +4,7 @@ import { CommonHooks } from '../../../../../../data';
 import SettingsBlock from '../../../../../ReusableComponents/SettingsBlock';
 import { Title } from '../../../../../ReusableComponents/Elements';
 
-const HooksTableBlock = () => {
+const HooksListBlock = () => {
 	const { webhooks } = CommonHooks.useWebhooks();
 	const { url, events } = webhooks;
 
@@ -46,4 +46,4 @@ const WebhookEvents = ( { events } ) => {
 	);
 };
 
-export default HooksTableBlock;
+export default HooksListBlock;
