+++ modules/ppcp-settings/resources/js/Components/Screens/Settings/Components/Settings/Blocks/OtherSettings.js
@@ -1,9 +1,9 @@
 import { __ } from '@wordpress/i18n';
 
-import Accordion from '../../../../ReusableComponents/AccordionSection';
-import SettingsBlock from '../../../../ReusableComponents/SettingsBlock';
-import { ControlSelect } from '../../../../ReusableComponents/Controls';
-import { SettingsHooks } from '../../../../../data';
+import Accordion from '../../../../../ReusableComponents/AccordionSection';
+import SettingsBlock from '../../../../../ReusableComponents/SettingsBlock';
+import { ControlSelect } from '../../../../../ReusableComponents/Controls';
+import { SettingsHooks } from '../../../../../../data';
 
 const OtherSettings = () => {
 	const { disabledCards, setDisabledCards } = SettingsHooks.useSettings();
